# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BakeryBiz operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
...
  /inventory/items:
    get:
      operationId: listInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List recent orders
      parameters:
        - $ref: "#/components/parameters/Search"
        - name: status
          in: query
          schema:
            type: string
            enum: [completed, pending, preparing, ready, delivered, cancelled, refunded]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Complete a POS order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order or insufficient stock
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Record an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - $ref: "#/components/parameters/Search"
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      responses:
        "200":
          description: Suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Search:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, categoryId, categoryName, price, cost, stock, minStock, active]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        categoryId: { type: integer }
        categoryName: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        cost: { type: number }
        stock: { type: number }
        minStock: { type: number }
        unit: { type: string }
        active: { type: boolean }
        imageUrl: { type: ["string", "null"] }
    ProductInput:
      type: object
      required: [name, sku, categoryId, price, cost, stock, minStock]
      properties:
        name: { type: string, minLength: 1 }
        sku: { type: string, minLength: 1 }
        categoryId: { type: integer }
        description: { type: string }
        price: { type: number, minimum: 0 }
        cost: { type: number, minimum: 0 }
        stock: { type: number, minimum: 0 }
        minStock: { type: number, minimum: 0 }
        unit: { type: string }
        imageUrl: { type: string }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        price: { type: number, minimum: 0 }
        cost: { type: number, minimum: 0 }
        minStock: { type: number, minimum: 0 }
        active: { type: boolean }
...
    InventoryItem:
      type: object
      required: [id, name, sku, categoryName, stock, minStock, unit, status, value]
      properties:
        id: { type: integer }
        name: { type: string }
        sku: { type: string }
        categoryName: { type: string }
        stock: { type: number }
        minStock: { type: number }
        unit: { type: string }
        status:
          type: string
          enum: [healthy, low, out]
        value: { type: number }
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId: { type: integer }
        quantity: { type: number, minimum: 0.01 }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, lineTotal]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        lineTotal: { type: number }
    Order:
      type: object
      required: [id, orderNumber, status, channel, paymentMethod, subtotal, discount, tax, total, createdAt, items]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        status: { type: string }
        channel: { type: string }
        paymentMethod: { type: string }
        customerId: { type: ["integer", "null"] }
        customerName: { type: ["string", "null"] }
        subtotal: { type: number }
        discount: { type: number }
        tax: { type: number }
        total: { type: number }
        createdAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
    OrderInput:
      type: object
      required: [items, paymentMethod]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/OrderItemInput" }
        paymentMethod: { type: string }
        customerId: { type: integer }
        discount: { type: number, minimum: 0 }
        tax: { type: number, minimum: 0 }
        channel: { type: string }
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
    Expense:
      type: object
      required: [id, category, description, amount, paymentMethod, expenseDate]
      properties:
        id: { type: integer }
        category: { type: string }
        description: { type: string }
        amount: { type: number }
        paymentMethod: { type: string }
        expenseDate: { type: string, format: date }
        vendor: { type: ["string", "null"] }
    ExpenseInput:
      type: object
      required: [category, description, amount, paymentMethod, expenseDate]
      properties:
        category: { type: string, minLength: 1 }
        description: { type: string, minLength: 1 }
        amount: { type: number, exclusiveMinimum: 0 }
        paymentMethod: { type: string }
        expenseDate: { type: string, format: date }
        vendor: { type: string }
    Customer:
      type: object
      required: [id, name, phone, totalOrders, totalSpent, outstandingCredit]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        totalOrders: { type: integer }
        totalSpent: { type: number }
        outstandingCredit: { type: number }
    CustomerInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 5 }
        email: { type: string }
    Supplier:
      type: object
      required: [id, name, phone, totalPurchases, outstanding]
      properties:
        id: { type: integer }
        name: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        totalPurchases: { type: number }
        outstanding: { type: number }
    SupplierInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string, minLength: 5 }
        email: { type: string }
...