# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BakeryBiz operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  id: number;
  name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  totalOrders: number;
  totalSpent: number;
  outstandingCredit: number;
}