# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BakeryBiz operations API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  category: string;
  description: string;
  amount: number;
  paymentMethod: string;
  expenseDate: Date;
  /** @nullable */
  vendor?: string | null;
}