# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BakeryBiz operations API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  description: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  paymentMethod: string;
  expenseDate: Date;
  vendor?: string;
}